Rename brain module to pipeline and update imports; run ruff
