+++ mindtrace/cluster/mindtrace/cluster/__init__.py
@@ -1,5 +1,5 @@
-from mindtrace.cluster.core.pipeline_worker import PipelineWorker
 from mindtrace.cluster.core.cluster import ClusterManager, Node, StandardWorkerLauncher, Worker
+from mindtrace.cluster.core.pipeline_worker import PipelineWorker
 from mindtrace.cluster.core.types import ProxyWorker
 
 __all__ = ["PipelineWorker", "ClusterManager", "Node", "Worker", "ProxyWorker", "StandardWorkerLauncher"]
