+++ tests/unit/mindtrace/cluster/test_pipeline_worker.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel
 
 from mindtrace.cluster import PipelineWorker
 from mindtrace.cluster.core.types import JobStatusEnum
-from mindtrace.models import Pipeline, BrainLoadInput, BrainUnloadInput
+from mindtrace.models import BrainLoadInput, BrainUnloadInput, Pipeline
 
 
 class EchoInput(BaseModel):
